code cleanup only; used eclipse cleanup to remove unnecessary code + organize imports.
git-svn-id: svn://cserg0.site.uottawa.ca/projetseg/trunk@1097 1d5a7336-c65e-6947-a78c-d4236001c669

+++ src/seg/jUCMNav/editparts/PathNodeEditPart.java
@@ -413,7 +413,7 @@ public class PathNodeEditPart extends ModelElementEditPart implements NodeEditPa
                         r = new Ray(list.getFirstPoint(), list.getMidpoint());
                     }
 
-                    double angle = Math.atan2((double) r.y, (double) r.x);
+                    double angle = Math.atan2(r.y, r.x);
 
                     ((Rotateable) nodeFigure).rotate(angle - Math.PI);
                 }
@@ -482,7 +482,7 @@ public class PathNodeEditPart extends ModelElementEditPart implements NodeEditPa
                         r = new Ray(list.getMidpoint(), list.getLastPoint());
                     }
 
-                    double angle = Math.atan2((double) r.y, (double) r.x);
+                    double angle = Math.atan2(r.y, r.x);
 
                     ((Rotateable) nodeFigure).rotate(angle - Math.PI);
                 }
